backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
Pillow==10.4.0
python-decouple==3.8
# psycopg2-binary==2.9.9  # Optional - only needed for PostgreSQL. Uncomment if using PostgreSQL
